WP VIP coding standard fixes

+++ laterpay/uninstall.php
@@ -5,27 +5,34 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
     exit;
 }
 
+if ( defined( 'WPCOM_IS_VIP_ENV' ) && true === WPCOM_IS_VIP_ENV ) {
+    return;
+}
+
 global $wpdb;
+
+// Check for non-VIP env.
 
 $table_terms_price   = $wpdb->prefix . 'laterpay_terms_price';
 $table_history       = $wpdb->prefix . 'laterpay_payment_history';
 $table_post_views    = $wpdb->prefix . 'laterpay_post_views';
 $table_time_passes   = $wpdb->prefix . 'laterpay_passes';
 $table_subscriptions = $wpdb->prefix . 'laterpay_subscriptions';
-$table_postmeta      = $wpdb->postmeta;
-$table_usermeta      = $wpdb->usermeta;
 
 // remove custom tables
 $sql = "
-    DROP TABLE IF EXISTS
-        $table_terms_price,
-        $table_history,
-        $table_post_views,
-        $table_time_passes,
-        $table_subscriptions
-    ;
+DROP TABLE IF EXISTS
+    $table_terms_price,
+    $table_history,
+    $table_post_views,
+    $table_time_passes,
+    $table_subscriptions
+;
 ";
 $wpdb->query( $sql );
+
+$table_postmeta      = $wpdb->postmeta;
+$table_usermeta      = $wpdb->usermeta;
 
 // remove pricing and voting data from wp_postmeta table
 delete_post_meta_by_key( 'laterpay_post_prices' );
@@ -80,11 +87,6 @@ delete_option( 'laterpay_preview_excerpt_word_count_max' );
 
 delete_option( 'laterpay_unlimited_access' );
 
-delete_option( 'laterpay_bulk_operations' );
-
-delete_option( 'laterpay_ratings' );
-delete_option( 'laterpay_hide_free_posts' );
-
 delete_option( 'laterpay_voucher_codes' );
 delete_option( 'laterpay_gift_codes' );
 delete_option( 'laterpay_voucher_statistic' );
@@ -95,9 +97,6 @@ delete_option( 'laterpay_debugger_addresses' );
 
 delete_option( 'laterpay_purchase_button_positioned_manually' );
 delete_option( 'laterpay_time_passes_positioned_manually' );
-delete_option( 'laterpay_show_time_passes_widget_on_free_posts' );
-
-delete_option( 'laterpay_landing_page' );
 
 delete_option( 'laterpay_only_time_pass_purchases_allowed' );
